requirements.txt
# Core Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# LLM Integration
google-generativeai>=0.3.0

# GitHub Integration
PyGithub>=2.1.1
requests>=2.31.0

# Database
sqlalchemy>=2.0.0
aiosqlite>=0.19.0

# Utilities
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
httpx>=0.25.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
